Developer Index

This page lists the types and functions that are internal to the ClusterValidityIndices.jl package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

Index

This section contains a list of internal names that link to their corresponding Documentation.

Methods

ClusterValidityIndices.get_internal_label!

ClusterValidityIndices.setup!

Types

ClusterValidityIndices.LabelMap

Constants

Docs

Documentation for all internal names are listed below.

ClusterValidityIndices.LabelMap — Type

Internal label mapping for incremental CVIs.

Alias for a dictionary mapping of integers to integers as cluster labels.

ClusterValidityIndices.get_internal_label! — Method

get_internal_label!(label_map, label)

Summary

get_internal_label!(label_map::Dict{Int64, Int64}, label::Int64) -> Int64

Get the internal label and update the label map if the label is new.

Arguments

label_map::LabelMap: label map to extract the internal label from.

label::Int: the external label that corresponds to an internal label.

Method List / Definition Locations

get_internal_label!(label_map, label)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/common.jl:141.

ClusterValidityIndices.setup! — Method

Summary

Internal method, sets up the CVI based upon the type of the provided sample.

Arguments

cvi::CVI: the CVI to setup to the correct dimensions.

sample::Vector{T<:RealFP}: The sample to use as a basis for setting up the CVI.

Method List / Definition Locations

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/CH.jl:79.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/DB.jl:76.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/GD43.jl:82.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/GD53.jl:82.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/PS.jl:67.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/WB.jl:84.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/XB.jl:76.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/cSIL.jl:72.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/rCIP.jl:82.