Index

ClusterValidityIndices.CH — Type

CH

The stateful information of the Calinski-Harabasz (CH) Cluster Validity Index

References

L. E. Brito da Silva, N. M. Melton, and D. C. Wunsch II, "Incremental Cluster Validity Indices for Hard Partitions: Extensions and Comparative Study," ArXiv e-prints, Feb 2019, arXiv:1902.06711v1 [cs.LG].

T. Calinski and J. Harabasz, "A dendrite method for cluster analysis," Communications in Statistics, vol. 3, no. 1, pp. 1–27, 1974.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML]. [Online].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1–23, 2018.

ClusterValidityIndices.CH — Method

CH()

Default constructor for the Calinski-Harabasz (CH) Cluster Validity Index.

ClusterValidityIndices.DB — Type

DB

The stateful information of the Davies-Bouldin (DB) Cluster Validity Index.

References

D. L. Davies and D. W. Bouldin, "A cluster separation measure," IEEE Transaction on Pattern Analysis and Machine Intelligence, vol. 1, no. 2, pp. 224–227, Feb. 1979.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML]. [Online].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1–23, 2018.

ClusterValidityIndices.DB — Method

DB()

Default constructor for the Davies-Bouldin (DB) Cluster Validity Index.

ClusterValidityIndices.GD43 — Type

GD43

The stateful information of the Generalized Dunn's Index 43 (GD43) Cluster Validity Index.

References

A. Ibrahim, J. M. Keller, and J. C. Bezdek, "Evaluating Evolving Structure in Streaming Data With Modified Dunn's Indices," IEEE Transactions on Emerging Topics in Computational Intelligence, pp. 1–12, 2019.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1–23, 2018.

J. C. Dunn, "A fuzzy relative of the ISODATA process and its use in detecting compact well-separated clusters," J. Cybern., vol. 3, no. 3 , pp. 32–57, 1973.

J. C. Bezdek and N. R. Pal, "Some new indexes of cluster validity," IEEE Trans. Syst., Man, and Cybern., vol. 28, no. 3, pp. 301–315, Jun. 1998.

ClusterValidityIndices.GD43 — Method

GD43()

Default constructor for the Generalized Dunn's Index 43 (GD43) Cluster Validity Index.

ClusterValidityIndices.GD53 — Type

GD53

The stateful information of the Generalized Dunn's Index 53 (GD53) Cluster Validity Index.

References

A. Ibrahim, J. M. Keller, and J. C. Bezdek, "Evaluating Evolving Structure in Streaming Data With Modified Dunn's Indices," IEEE Transactions on Emerging Topics in Computational Intelligence, pp. 1–12, 2019.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1–23, 2018.

J. C. Dunn, "A fuzzy relative of the ISODATA process and its use in detecting compact well-separated clusters," J. Cybern., vol. 3, no. 3 , pp. 32–57, 1973.

J. C. Bezdek and N. R. Pal, "Some new indexes of cluster validity," IEEE Trans. Syst., Man, and Cybern., vol. 28, no. 3, pp. 301–315, Jun. 1998.

ClusterValidityIndices.GD53 — Method

GD53()

Default constructor for the Generalized Dunn's Index 53 (GD53) Cluster Validity Index.

ClusterValidityIndices.PS — Type

PS

The stateful information of the Partition Separation (PS) Cluster Validity Index.

References

Miin-Shen Yang and Kuo-Lung Wu, "A new validity index for fuzzy clustering," 10th IEEE International Conference on Fuzzy Systems. (Cat. No.01CH37297), Melbourne, Victoria, Australia, 2001, pp. 89-92, vol.1.

E. Lughofer, "Extensions of vector quantization for incremental clustering," Pattern Recognit., vol. 41, no. 3, pp. 995–1011, 2008.

ClusterValidityIndices.PS — Method

PS()

Default constructor for the Partition Separation (PS) Cluster Validity Index.

ClusterValidityIndices.WB — Type

WB

The stateful information of the WB-Index (WB) Cluster Validity Index.

References

L. E. Brito da Silva, N. M. Melton, and D. C. Wunsch II, "Incremental Cluster Validity Indices for Hard Partitions: Extensions and Comparative Study," ArXiv e-prints, Feb 2019, arXiv:1902.06711v1 [cs.LG].

Q. Zhao, M. Xu, and P. Franti, "Sum-of-Squares Based Cluster Validity Index and Significance Analysis," in Adaptive and Natural Computing Algorithms, M. Kolehmainen, P. Toivanen, and B. Beliczynski, Eds. Berlin, Heidelberg: Springer Berlin Heidelberg, 2009, pp. 313–322.

Q. Zhao and P. Franti, "WB-index: A sum-of-squares based index for cluster validity," Data Knowledge Engineering, vol. 92, pp. 77–89, 2014.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1–23, 2018.

ClusterValidityIndices.WB — Method

WB()

Default constructor for the WB-Index (WB) Cluster Validity Index.

ClusterValidityIndices.XB — Type

XB

The stateful information of the Xie-Beni (XB) Cluster Validity Index.

References

X. L. Xie and G. Beni, "A Validity Measure for Fuzzy Clustering," IEEE Transactions on Pattern Analysis and Machine Intelligence, vol. 13, no. 8, pp. 841–847, 1991.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML]. [Online].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1–23, 2018.

ClusterValidityIndices.XB — Method

XB()

Default constructor for the Xie-Beni (XB) Cluster Validity Index.

ClusterValidityIndices.cSIL — Type

cSIL

The stateful information of the Centroid-based Silhouette (cSIL) Cluster Validity Index.

References

L. E. Brito da Silva, N. M. Melton, and D. C. Wunsch II, "Incremental Cluster Validity Indices for Hard Partitions: Extensions and Comparative Study," ArXiv e-prints, Feb 2019, arXiv:1902.06711v1 [cs.LG].

P. J. Rousseeuw, "Silhouettes: A graphical aid to the interpretation and validation of cluster analysis," Journal of Computational and Applied Mathematics, vol. 20, pp. 53–65, 1987.

M. Rawashdeh and A. Ralescu, "Center-wise intra-inter silhouettes," in Scalable Uncertainty Management, E. Hüllermeier, S. Link, T. Fober et al., Eds. Berlin, Heidelberg: Springer, 2012, pp. 406–419.

ClusterValidityIndices.cSIL — Method

cSIL()

Default constructor for the Centroid-based Silhouette (cSIL) Cluster Validity Index.

ClusterValidityIndices.rCIP — Type

rCIP

The stateful information of the (Renyi's) representative Cross Information Potential (rCIP) Cluster Validity Index.

References

L. E. Brito da Silva, N. M. Melton, and D. C. Wunsch II, "Incremental Cluster Validity Indices for Hard Partitions: Extensions and Comparative Study," ArXiv e-prints, Feb 2019, arXiv:1902.06711v1 [cs.LG].

E. Gokcay and J. C. Principe, "A new clustering evaluation function using Renyi’s information potential," in Proc. Int. Conf. Acoust., Speech, Signal Process. (ICASSP), vol. 6. Jun. 2000, pp. 3490–3493.

E. Gokcay and J. C. Principe, "Information theoretic clustering," IEEE Trans. Pattern Anal. Mach. Intell., vol. 24, no. 2, pp. 158–171, Feb. 2002.

D. Araújo, A. D. Neto, and A. Martins, "Representative cross information potential clustering," Pattern Recognit. Lett., vol. 34, no. 16, pp. 2181–2191, Dec. 2013.

D. Araújo, A. D. Neto, and A. Martins, "Information-theoretic clustering: A representative and evolutionary approach," Expert Syst. Appl., vol. 40, no. 10, pp. 4190–4205, Aug. 2013.

R. O. Duda, P. E. Hart, and D. G. Stork, Pattern Classification, 2nd ed. John Wiley & Sons, 2000.

ClusterValidityIndices.rCIP — Method

rCIP()

Default constructor for the (Renyi's) representative Cross Information Potential (rCIP) Cluster Validity Index.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::CVI)

Compute the criterion value of the CVI.

After computation, the resulting criterion value can be extracted from cvi.criterion_value. The criterion value is a function of the CVI/ICVI internal parameters, so at least two classes (i.e., unique labels) must be presented to the CVI in param_inc! or param_batch! before a non-zero value is returned.

Arguments

cvi::CVI: the stateful information of the CVI/ICVI to use for computing the criterion value.

Examples

julia> my_cvi = CH()
julia> data = load_some_data()
julia> labels = my_cluster_algorithm(data)
julia> param_batch!(my_cvi, data, labels)
julia> evaluate!(my_cvi)
julia> my_criterion_value = my_cvi.criterion_value

ClusterValidityIndices.get_cvi! — Method

get_cvi!(cvi::CVI, data::RealMatrix, labels::IntegerVector)

Compute and return the criterion value in batch mode.

Arguments

cvi::CVI: the stateful information of the CVI providing the criterion value.

data::RealMatrix: a matrix of data, columns as samples and rows as features, used in the external clustering process.

labels::IntegerVector: a vector of integers representing labels prescribed to the data by the external clustering algorithm.

Examples

# Create a new CVI object
my_cvi = CH()
# Load in data from some external source
data = load_some_data()
# Cluster the data into a set of labels as an integer vector
labels = my_cluster_algorithm(data)
# Compute the final criterion value in batch mode
criterion_value = get_cvi!(cvi, data, labels)

ClusterValidityIndices.get_icvi! — Method

get_icvi!(cvi::CVI, sample::RealVector, label::Integer)

Compute and return the criterion value incrementally.

Arguments

cvi::CVI: the stateful information of the ICVI providing the criterion value.

sample::RealVector: the sample provided to the external. clustering algorithm.

label::Integer: the label prescribed to the sample by the external clustering algorithm.

Examples

# Create a new CVI object
my_cvi = CH()
# Load in data from some external source
data = load_some_data()
# Cluster the data into a set of labels as an integer vector
labels = my_cluster_algorithm(data)
# Iteratively compute and extract the criterion value at every step
n_samples = length(labels)
criterion_values = zeros(n_samples)
for ix = 1:n_samples
    sample = data[:, ix]
    label = labels[ix]
    criterion_values[ix] = get_icvi!(my_cvi, sample, label)
end

ClusterValidityIndices.get_internal_label! — Method

get_internal_label!(label_map::LabelMap, label::Int)

Get the internal label and update the label map if the label is new.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::CVI, data::RealMatrix, labels::IntegerVector)

Compute the CVI parameters in batch.

This method updates only the internal parameters of the CVI algorithm in batch. When the criterion value itself is needed, use evaluate! and extract it from cvi.criterion_value.

Arguments

cvi::CVI: the stateful information of the CVI/ICVI algorithm.

data::RealMatrix: a matrix of data where rows are features and columns are samples, used in the external clustering algorithm.

labels::IntegerVector: a vector of labels that the external clustering algorithm prescribed to each column in data.

Examples

julia> my_cvi = CH()
julia> data = load_some_data()
julia> labels = my_cluster_algorithm(data)
julia> param_batch!(my_cvi, data, labels)

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::CVI, sample::RealVector, label::Integer)

Compute the CVI parameters incrementally.

This method updates only internal parameters of the ICVI algorithm incrementally. When the criterion value itself is needed, use evaluate! and extract it from cvi.criterion_value.

Arguments

cvi::CVI: the stateful information of the CVI/ICVI algorithm.

sample::RealVector: a vector of features used in the external clustering algorithm.

label::Integer: the label that the external clustering algorithm prescribed to the sample.

Examples

julia> my_cvi = CH()
julia> data = load_some_data()
julia> labels = my_cluster_algorithm(data)
julia> param_inc!(my_cvi, data[:, 1], labels[1])

ClusterValidityIndices.relabel_cvi_data — Method

relabel_cvi_data(labels::IntegerVector)

Relabels the vector to present new labels in incremental order.

ClusterValidityIndices.setup! — Method

setup!(cvi::CH, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.setup! — Method

setup!(cvi::DB, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.setup! — Method

setup!(cvi::GD43, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.setup! — Method

setup!(cvi::GD53, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.setup! — Method

setup!(cvi::PS, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.setup! — Method

setup!(cvi::WB, sample::Vector{T}) where {T<:Real}

ClusterValidityIndices.setup! — Method

setup!(cvi::XB, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.setup! — Method

setup!(cvi::cSIL, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.setup! — Method

setup!(cvi::rCIP, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.sort_cvi_data — Method

sort_cvi_data(data::RealMatrix, labels::IntegerVector)

Sorts the CVI data by the label index, assuring that clusters are provided incrementally.