Index

This page lists the core methods and types of the ClusterValidityIndices.jl package. The Modules section lists the modules exported by the package including the ClusterValidityIndices module itself. The Methods section lists the public methods for the package that use the CVIs/ICVIs in Types. Each of these entries link to the docstrings in the Docs section.

Index

This section enumerates the names exported by the package, each of which links to its corresponding Documentation.

Modules

ClusterValidityIndices.ClusterValidityIndices

Functions

ClusterValidityIndices.get_cvi!

ClusterValidityIndices.get_cvi!

Types

ClusterValidityIndices.CH

ClusterValidityIndices.CH

ClusterValidityIndices.CVI

ClusterValidityIndices.DB

ClusterValidityIndices.DB

ClusterValidityIndices.GD43

ClusterValidityIndices.GD43

ClusterValidityIndices.GD53

ClusterValidityIndices.GD53

ClusterValidityIndices.PS

ClusterValidityIndices.PS

ClusterValidityIndices.WB

ClusterValidityIndices.WB

ClusterValidityIndices.XB

ClusterValidityIndices.XB

ClusterValidityIndices.cSIL

ClusterValidityIndices.cSIL

ClusterValidityIndices.rCIP

ClusterValidityIndices.rCIP

Constants

ClusterValidityIndices.CLUSTERVALIDITYINDICES_VERSION

ClusterValidityIndices.CVI_MODULES

Docs

This section lists the documentation for every exported name of the ClusterValidityIndices.jl package.

Modules

ClusterValidityIndices.ClusterValidityIndices — Module

Main module for ClusterValidityIndices.jl, a Julia package of metrics for unsupervised learning.

This module exports all of the CVI modules, options, and utilities used by the ClusterValidityIndices.jl package. For full usage, see the official guide at https://ap6yc.github.io/ClusterValidityIndices.jl/dev/man/guide/.

Basic Usage

Install and import the package in a script with

using Pkg
Pkg.add("ClusterValidityIndices")
using ClusterValidityIndices

then create a CVI object with an empty argument constructor

my_cvi = DB()

and get the criterion values with get_cvi! (batch) or get_icvi! (incremental)

# Load some features and labels from a clustering process
features, labels = get_some_clustering_data()

# Batch criterion value
criterion_value = get_cvi!(my_cvi, features, labels)

# Incremental criterion values
criterion_values = zeros(length(labels))
for ix in eachindex(labels)
    criterion_values[ix] = get_icvi!(my_cvi, features[:, ix], labels[ix])
end

Imports

The following names are imported by the package as dependencies:

Base

Core

DocStringExtensions

LinearAlgebra

NumericalTypeAliases

Pkg

Exports

The following names are exported and available when using the package:

CH

CLUSTERVALIDITYINDICES_VERSION

CVI

CVI_MODULES

DB

GD43

GD53

PS

WB

XB

cSIL

get_cvi!

rCIP

Functions

ClusterValidityIndices.get_cvi! — Method

get_cvi!(
    cvi::CVI,
    data::AbstractMatrix{T} where T<:Real,
    labels::AbstractVector{T} where T<:Integer
) -> Any

Summary

Compute and return the criterion value in batch mode.

This method takes the CVI object, a batch of samples as a matrix of floats, and a vector of integers that represent the labels prescribed to the data by your clustering algorithm.

Note

You cannot switch to incremental mode after evaluating a CVI in batch mode. To evaluate incrementally, you much create a new CVI object.

Arguments

cvi::CVI: the stateful information of the CVI providing the criterion value.

data::RealMatrix: a matrix of data, columns as samples and rows as features, used in the external clustering process.

labels::IntegerVector: a vector of integers representing labels prescribed to the data by the external clustering algorithm.

Examples

# Create a new CVI object
my_cvi = CH()

# Load in random data as an example; 10 samples with feature dimenison 3
dim = 3
n_samples = 10
data = rand(dim, n_samples)
labels = repeat(1:2, inner=n_samples)

# Compute the final criterion value in batch mode
criterion_value = get_cvi!(cvi, data, labels)

Method List / Definition Locations

get_cvi!(cvi, data, labels)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/common.jl:145.

ClusterValidityIndices.get_cvi! — Method

get_cvi!(
    cvi::CVI,
    sample::AbstractVector{T} where T<:Real,
    label::Integer
) -> Any

Summary

Compute and return the criterion value incrementally.

This method takes the CVI object, a single sample as a vector of floats, and a single integer that represents the label prescribed to the sample by your clustering algorithm.

Note

You cannot switch to batch mode after incrementally evaluating a CVI. To evaluate in batch, you much create a new CVI object.

Arguments

cvi::CVI: the stateful information of the ICVI providing the criterion value.

sample::RealVector: a vector of features used in clustering the sample.

label::Integer: the cluster label prescribed to the sample by the clustering algorithm.

Examples

# Create a new CVI object
my_cvi = CH()

# Load in random data as an example; 10 samples with feature dimenison 3
dim = 3
n_samples = 10
data = rand(dim, n_samples)
labels = repeat(1:2, inner=n_samples)

# Iteratively compute and extract the criterion value at every step
criterion_values = zeros(n_samples)
for ix = 1:n_samples
    sample = data[:, ix]
    label = labels[ix]
    criterion_values[ix] = get_icvi!(my_cvi, sample, label)
end

Method List / Definition Locations

get_cvi!(cvi, sample, label)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/common.jl:105.

Types

ClusterValidityIndices.CH — Type

mutable struct CH <: CVI

Summary

The stateful information of the Calinski-Harabasz (CH) Cluster Validity Index

References

L. E. Brito da Silva, N. M. Melton, and D. C. Wunsch II, "Incremental Cluster Validity Indices for Hard Partitions: Extensions and Comparative Study," ArXiv e-prints, Feb 2019, arXiv:1902.06711v1 [cs.LG].

T. Calinski and J. Harabasz, "A dendrite method for cluster analysis," Communications in Statistics, vol. 3, no. 1, pp. 1-27, 1974.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML]. [Online].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1-23, 2018.

Fields

label_map::Dict{Int64, Int64}

dim::Int64

n_samples::Int64

mu::Vector{Float64}

n::Vector{Int64}

v::Matrix{Float64}

CP::Vector{Float64}

SEP::Vector{Float64}

G::Matrix{Float64}

BGSS::Float64

WGSS::Float64

n_clusters::Int64

criterion_value::Float64

ClusterValidityIndices.CH — Method

CH() -> CH

Summary

Constructor for the Calinski-Harabasz (CH) Cluster Validity Index.

Examples

# Import the package
using ClusterValidityIndices
# Construct a CH module
my_cvi = CH()

References

L. E. Brito da Silva, N. M. Melton, and D. C. Wunsch II, "Incremental Cluster Validity Indices for Hard Partitions: Extensions and Comparative Study," ArXiv e-prints, Feb 2019, arXiv:1902.06711v1 [cs.LG].

T. Calinski and J. Harabasz, "A dendrite method for cluster analysis," Communications in Statistics, vol. 3, no. 1, pp. 1-27, 1974.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML]. [Online].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1-23, 2018.

Method List / Definition Locations

CH()

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/CH.jl:70.

ClusterValidityIndices.CVI — Type

abstract type CVI

Summary

Abstract supertype for all CVI objects. All index instantiations are subtypes of CVI.

Fields

ClusterValidityIndices.DB — Type

mutable struct DB <: CVI

Summary

The stateful information of the Davies-Bouldin (DB) Cluster Validity Index.

References

D. L. Davies and D. W. Bouldin, "A cluster separation measure," IEEE Transaction on Pattern Analysis and Machine Intelligence, vol. 1, no. 2, pp. 224-227, Feb. 1979.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML]. [Online].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1-23, 2018.

Fields

label_map::Dict{Int64, Int64}

dim::Int64

n_samples::Int64

mu_data::Vector{Float64}

n::Vector{Int64}

v::Matrix{Float64}

CP::Vector{Float64}

S::Vector{Float64}

R::Matrix{Float64}

G::Matrix{Float64}

D::Matrix{Float64}

n_clusters::Int64

criterion_value::Float64

ClusterValidityIndices.DB — Method

DB() -> DB

Summary

Constructor for the Davies-Bouldin (DB) Cluster Validity Index.

Examples

# Import the package
using ClusterValidityIndices
# Construct a DB module
my_cvi = DB()

References

D. L. Davies and D. W. Bouldin, "A cluster separation measure," IEEE Transaction on Pattern Analysis and Machine Intelligence, vol. 1, no. 2, pp. 224-227, Feb. 1979.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML]. [Online].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1-23, 2018.

Method List / Definition Locations

DB()

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/DB.jl:67.

ClusterValidityIndices.GD43 — Type

mutable struct GD43 <: CVI

Summary

The stateful information of the Generalized Dunn's Index 43 (GD43) Cluster Validity Index.

References

A. Ibrahim, J. M. Keller, and J. C. Bezdek, "Evaluating Evolving Structure in Streaming Data With Modified Dunn's Indices," IEEE Transactions on Emerging Topics in Computational Intelligence, pp. 1-12, 2019.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1-23, 2018.

J. C. Dunn, "A fuzzy relative of the ISODATA process and its use in detecting compact well-separated clusters," J. Cybern., vol. 3, no. 3 , pp. 32-57, 1973.

J. C. Bezdek and N. R. Pal, "Some new indexes of cluster validity," IEEE Trans. Syst., Man, and Cybern., vol. 28, no. 3, pp. 301-315, Jun. 1998.

Fields

label_map::Dict{Int64, Int64}

dim::Int64

n_samples::Int64

mu_data::Vector{Float64}

n::Vector{Int64}

v::Matrix{Float64}

CP::Vector{Float64}

G::Matrix{Float64}

D::Matrix{Float64}

inter::Float64

intra::Float64

n_clusters::Int64

criterion_value::Float64

ClusterValidityIndices.GD43 — Method

GD43() -> GD43

Summary

Constructor for the Generalized Dunn's Index 43 (GD43) Cluster Validity Index.

Examples

# Import the package
using ClusterValidityIndices
# Construct a GD43 module
my_cvi = GD43()

References

A. Ibrahim, J. M. Keller, and J. C. Bezdek, "Evaluating Evolving Structure in Streaming Data With Modified Dunn's Indices," IEEE Transactions on Emerging Topics in Computational Intelligence, pp. 1-12, 2019.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1-23, 2018.

J. C. Dunn, "A fuzzy relative of the ISODATA process and its use in detecting compact well-separated clusters," J. Cybern., vol. 3, no. 3 , pp. 32-57, 1973.

J. C. Bezdek and N. R. Pal, "Some new indexes of cluster validity," IEEE Trans. Syst., Man, and Cybern., vol. 28, no. 3, pp. 301-315, Jun. 1998.

Method List / Definition Locations

GD43()

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/GD43.jl:73.

ClusterValidityIndices.GD53 — Type

mutable struct GD53 <: CVI

Summary

The stateful information of the Generalized Dunn's Index 53 (GD53) Cluster Validity Index.

References

A. Ibrahim, J. M. Keller, and J. C. Bezdek, "Evaluating Evolving Structure in Streaming Data With Modified Dunn's Indices," IEEE Transactions on Emerging Topics in Computational Intelligence, pp. 1-12, 2019.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1-23, 2018.

J. C. Dunn, "A fuzzy relative of the ISODATA process and its use in detecting compact well-separated clusters," J. Cybern., vol. 3, no. 3 , pp. 32-57, 1973.

J. C. Bezdek and N. R. Pal, "Some new indexes of cluster validity," IEEE Trans. Syst., Man, and Cybern., vol. 28, no. 3, pp. 301-315, Jun. 1998.

Fields

label_map::Dict{Int64, Int64}

dim::Int64

n_samples::Int64

mu_data::Vector{Float64}

n::Vector{Int64}

v::Matrix{Float64}

CP::Vector{Float64}

G::Matrix{Float64}

D::Matrix{Float64}

inter::Float64

intra::Float64

n_clusters::Int64

criterion_value::Float64

ClusterValidityIndices.GD53 — Method

GD53() -> GD53

Summary

Constructor for the Generalized Dunn's Index 53 (GD53) Cluster Validity Index.

Examples

# Import the package
using ClusterValidityIndices
# Construct a GD53 module
my_cvi = GD53()

References

A. Ibrahim, J. M. Keller, and J. C. Bezdek, "Evaluating Evolving Structure in Streaming Data With Modified Dunn's Indices," IEEE Transactions on Emerging Topics in Computational Intelligence, pp. 1-12, 2019.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1-23, 2018.

J. C. Dunn, "A fuzzy relative of the ISODATA process and its use in detecting compact well-separated clusters," J. Cybern., vol. 3, no. 3 , pp. 32-57, 1973.

J. C. Bezdek and N. R. Pal, "Some new indexes of cluster validity," IEEE Trans. Syst., Man, and Cybern., vol. 28, no. 3, pp. 301-315, Jun. 1998.

Method List / Definition Locations

GD53()

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/GD53.jl:73.

ClusterValidityIndices.PS — Type

mutable struct PS <: CVI

Summary

The stateful information of the Partition Separation (PS) Cluster Validity Index.

References

Miin-Shen Yang and Kuo-Lung Wu, "A new validity index for fuzzy clustering," 10th IEEE International Conference on Fuzzy Systems. (Cat. No.01CH37297), Melbourne, Victoria, Australia, 2001, pp. 89-92, vol.1.

E. Lughofer, "Extensions of vector quantization for incremental clustering," Pattern Recognit., vol. 41, no. 3, pp. 995-1011, 2008.

Fields

label_map::Dict{Int64, Int64}

dim::Int64

n_samples::Int64

n::Vector{Int64}

v::Matrix{Float64}

D::Matrix{Float64}

v_bar::Vector{Float64}

beta_t::Float64

PS_i::Vector{Float64}

n_clusters::Int64

criterion_value::Float64

ClusterValidityIndices.PS — Method

PS() -> PS

Summary

Constructor for the Partition Separation (PS) Cluster Validity Index.

Examples

# Import the package
using ClusterValidityIndices
# Construct a PS module
my_cvi = PS()

References

Miin-Shen Yang and Kuo-Lung Wu, "A new validity index for fuzzy clustering," 10th IEEE International Conference on Fuzzy Systems. (Cat. No.01CH37297), Melbourne, Victoria, Australia, 2001, pp. 89-92, vol.1.

E. Lughofer, "Extensions of vector quantization for incremental clustering," Pattern Recognit., vol. 41, no. 3, pp. 995-1011, 2008.

Method List / Definition Locations

PS()

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/PS.jl:60.

ClusterValidityIndices.WB — Type

mutable struct WB <: CVI

Summary

The stateful information of the WB-Index (WB) Cluster Validity Index.

References

L. E. Brito da Silva, N. M. Melton, and D. C. Wunsch II, "Incremental Cluster Validity Indices for Hard Partitions: Extensions and Comparative Study," ArXiv e-prints, Feb 2019, arXiv:1902.06711v1 [cs.LG].

Q. Zhao, M. Xu, and P. Franti, "Sum-of-Squares Based Cluster Validity Index and Significance Analysis," in Adaptive and Natural Computing Algorithms, M. Kolehmainen, P. Toivanen, and B. Beliczynski, Eds. Berlin, Heidelberg: Springer Berlin Heidelberg, 2009, pp. 313-322.

Q. Zhao and P. Franti, "WB-index: A sum-of-squares based index for cluster validity," Data Knowledge Engineering, vol. 92, pp. 77-89, 2014.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1-23, 2018.

Fields

label_map::Dict{Int64, Int64}

dim::Int64

n_samples::Int64

mu::Vector{Float64}

n::Vector{Int64}

v::Matrix{Float64}

CP::Vector{Float64}

SEP::Vector{Float64}

G::Matrix{Float64}

BGSS::Float64

WGSS::Float64

n_clusters::Int64

criterion_value::Float64

ClusterValidityIndices.WB — Method

WB() -> WB

Summary

Constructor for the WB-Index (WB) Cluster Validity Index.

Examples

# Import the package
using ClusterValidityIndices
# Construct a WB module
my_cvi = WB()

References

L. E. Brito da Silva, N. M. Melton, and D. C. Wunsch II, "Incremental Cluster Validity Indices for Hard Partitions: Extensions and Comparative Study," ArXiv e-prints, Feb 2019, arXiv:1902.06711v1 [cs.LG].

Q. Zhao, M. Xu, and P. Franti, "Sum-of-Squares Based Cluster Validity Index and Significance Analysis," in Adaptive and Natural Computing Algorithms, M. Kolehmainen, P. Toivanen, and B. Beliczynski, Eds. Berlin, Heidelberg: Springer Berlin Heidelberg, 2009, pp. 313-322.

Q. Zhao and P. Franti, "WB-index: A sum-of-squares based index for cluster validity," Data Knowledge Engineering, vol. 92, pp. 77-89, 2014.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1-23, 2018.

Method List / Definition Locations

WB()

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/WB.jl:75.

ClusterValidityIndices.XB — Type

mutable struct XB <: CVI

Summary

The stateful information of the Xie-Beni (XB) Cluster Validity Index.

References

X. L. Xie and G. Beni, "A Validity Measure for Fuzzy Clustering," IEEE Transactions on Pattern Analysis and Machine Intelligence, vol. 13, no. 8, pp. 841-847, 1991.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML]. [Online].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1-23, 2018.

Fields

label_map::Dict{Int64, Int64}

dim::Int64

n_samples::Int64

mu_data::Vector{Float64}

n::Vector{Int64}

v::Matrix{Float64}

CP::Vector{Float64}

SEP::Float64

G::Matrix{Float64}

D::Matrix{Float64}

WGSS::Float64

n_clusters::Int64

criterion_value::Float64

ClusterValidityIndices.XB — Method

XB() -> XB

Summary

Constructor for the Xie-Beni (XB) Cluster Validity Index.

Examples

# Import the package
using ClusterValidityIndices
# Construct a XB module
my_cvi = XB()

References

X. L. Xie and G. Beni, "A Validity Measure for Fuzzy Clustering," IEEE Transactions on Pattern Analysis and Machine Intelligence, vol. 13, no. 8, pp. 841-847, 1991.

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML]. [Online].

M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1-23, 2018.

Method List / Definition Locations

XB()

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/XB.jl:67.

ClusterValidityIndices.cSIL — Type

mutable struct cSIL <: CVI

Summary

The stateful information of the Centroid-based Silhouette (cSIL) Cluster Validity Index.

References

L. E. Brito da Silva, N. M. Melton, and D. C. Wunsch II, "Incremental Cluster Validity Indices for Hard Partitions: Extensions and Comparative Study," ArXiv e-prints, Feb 2019, arXiv:1902.06711v1 [cs.LG].

P. J. Rousseeuw, "Silhouettes: A graphical aid to the interpretation and validation of cluster analysis," Journal of Computational and Applied Mathematics, vol. 20, pp. 53-65, 1987.

M. Rawashdeh and A. Ralescu, "Center-wise intra-inter silhouettes," in Scalable Uncertainty Management, E. Hüllermeier, S. Link, T. Fober et al., Eds. Berlin, Heidelberg: Springer, 2012, pp. 406-419.

Fields

label_map::Dict{Int64, Int64}

dim::Int64

n_samples::Int64

n::Vector{Int64}

v::Matrix{Float64}

CP::Vector{Float64}

G::Matrix{Float64}

S::Matrix{Float64}

sil_coefs::Vector{Float64}

n_clusters::Int64

criterion_value::Float64

ClusterValidityIndices.cSIL — Method

cSIL() -> cSIL

Summary

Constructor for the Centroid-based Silhouette (cSIL) Cluster Validity Index.

Examples

# Import the package
using ClusterValidityIndices
# Construct a cSIL module
my_cvi = cSIL()

References

L. E. Brito da Silva, N. M. Melton, and D. C. Wunsch II, "Incremental Cluster Validity Indices for Hard Partitions: Extensions and Comparative Study," ArXiv e-prints, Feb 2019, arXiv:1902.06711v1 [cs.LG].

P. J. Rousseeuw, "Silhouettes: A graphical aid to the interpretation and validation of cluster analysis," Journal of Computational and Applied Mathematics, vol. 20, pp. 53-65, 1987.

M. Rawashdeh and A. Ralescu, "Center-wise intra-inter silhouettes," in Scalable Uncertainty Management, E. Hüllermeier, S. Link, T. Fober et al., Eds. Berlin, Heidelberg: Springer, 2012, pp. 406-419.

Method List / Definition Locations

cSIL()

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/cSIL.jl:65.

ClusterValidityIndices.rCIP — Type

mutable struct rCIP <: CVI

Summary

The stateful information of the (Renyi's) representative Cross Information Potential (rCIP) Cluster Validity Index.

References

L. E. Brito da Silva, N. M. Melton, and D. C. Wunsch II, "Incremental Cluster Validity Indices for Hard Partitions: Extensions and Comparative Study," ArXiv e-prints, Feb 2019, arXiv:1902.06711v1 [cs.LG].

E. Gokcay and J. C. Principe, "A new clustering evaluation function using Renyi's information potential," in Proc. Int. Conf. Acoust., Speech, Signal Process. (ICASSP), vol. 6. Jun. 2000, pp. 3490-3493.

E. Gokcay and J. C. Principe, "Information theoretic clustering," IEEE Trans. Pattern Anal. Mach. Intell., vol. 24, no. 2, pp. 158-171, Feb. 2002.

D. Araújo, A. D. Neto, and A. Martins, "Representative cross information potential clustering," Pattern Recognit. Lett., vol. 34, no. 16, pp. 2181-2191, Dec. 2013.

D. Araújo, A. D. Neto, and A. Martins, "Information-theoretic clustering: A representative and evolutionary approach," Expert Syst. Appl., vol. 40, no. 10, pp. 4190-4205, Aug. 2013.

R. O. Duda, P. E. Hart, and D. G. Stork, Pattern Classification, 2nd ed. John Wiley & Sons, 2000.

Fields

label_map::Dict{Int64, Int64}

dim::Int64

n_samples::Int64

n::Vector{Int64}

v::Matrix{Float64}

sigma::Array{Float64, 3}

constant::Float64

D::Matrix{Float64}

delta_term::Matrix{Float64}

n_clusters::Int64

criterion_value::Float64

ClusterValidityIndices.rCIP — Method

rCIP() -> rCIP

Summary

Constructor for the (Renyi's) representative Cross Information Potential (rCIP) Cluster Validity Index.

Examples

# Import the package
using ClusterValidityIndices
# Construct a rCIP module
my_cvi = rCIP()

References

L. E. Brito da Silva, N. M. Melton, and D. C. Wunsch II, "Incremental Cluster Validity Indices for Hard Partitions: Extensions and Comparative Study," ArXiv e-prints, Feb 2019, arXiv:1902.06711v1 [cs.LG].

E. Gokcay and J. C. Principe, "A new clustering evaluation function using Renyi's information potential," in Proc. Int. Conf. Acoust., Speech, Signal Process. (ICASSP), vol. 6. Jun. 2000, pp. 3490-3493.

E. Gokcay and J. C. Principe, "Information theoretic clustering," IEEE Trans. Pattern Anal. Mach. Intell., vol. 24, no. 2, pp. 158-171, Feb. 2002.

D. Araújo, A. D. Neto, and A. Martins, "Representative cross information potential clustering," Pattern Recognit. Lett., vol. 34, no. 16, pp. 2181-2191, Dec. 2013.

D. Araújo, A. D. Neto, and A. Martins, "Information-theoretic clustering: A representative and evolutionary approach," Expert Syst. Appl., vol. 40, no. 10, pp. 4190-4205, Aug. 2013.

R. O. Duda, P. E. Hart, and D. G. Stork, Pattern Classification, 2nd ed. John Wiley & Sons, 2000.

Method List / Definition Locations

rCIP()

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/rCIP.jl:75.

Constants

ClusterValidityIndices.CLUSTERVALIDITYINDICES_VERSION — Constant

CLUSTERVALIDITYINDICES_VERSION

Description

A constant that contains the version of the installed ClusterValidityIndices.jl package.

This value is computed at compile time, so it may be used to programmatically verify the version of ClusterValidityIndices that is installed in case a compat entry in your Project.toml is missing or otherwise incorrect.

ClusterValidityIndices.CVI_MODULES — Constant

CVI_MODULES

Description

List of implemented CVIs, useful for iteration. Each element is the struct abbreviated name for the CVI, which can be instantiated for iteration with the empty constructor.

For example:

using ClusterValidityIndices
instantiated_cvis = [local_cvi() for local_cvi in CVI_MODULES]