Developer Index

This page lists the types and functions that are internal to the ClusterValidityIndices.jl package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

Index

This section contains a list of internal names that link to their corresponding Documentation.

Methods

ClusterValidityIndices.evaluate!

ClusterValidityIndices.get_internal_label!

ClusterValidityIndices.param_batch!

ClusterValidityIndices.param_inc!

ClusterValidityIndices.setup!

Types

ClusterValidityIndices.LabelMap

Constants

Docs

Documentation for all internal names are listed below.

ClusterValidityIndices.LabelMap — Type

LabelMap

Description

Internal label mapping for incremental CVIs.

Alias for a dictionary mapping of integers to integers as cluster labels.

ClusterValidityIndices.evaluate! — Method

Summary

Compute the criterion value of the CVI.

After computation, the resulting criterion value can be extracted from cvi.criterion_value. The criterion value is a function of the CVI/ICVI internal parameters, so at least two classes (i.e., unique labels) must be presented to the CVI in param_inc! or param_batch! before a non-zero value is returned.

Arguments

cvi::CVI: the stateful information of the CVI/ICVI to use for computing the criterion value.

Examples

julia> my_cvi = CH()
julia> data = load_some_data()
julia> labels = my_cluster_algorithm(data)
julia> param_batch!(my_cvi, data, labels)
julia> evaluate!(my_cvi)
julia> my_criterion_value = my_cvi.criterion_value

Method List / Definition Locations

evaluate!(cvi)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/CH.jl:179.

evaluate!(cvi)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/DB.jl:209.

evaluate!(cvi)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/GD43.jl:211.

evaluate!(cvi)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/GD53.jl:209.

evaluate!(cvi)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/PS.jl:160.

evaluate!(cvi)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/WB.jl:186.

evaluate!(cvi)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/XB.jl:204.

evaluate!(cvi)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/cSIL.jl:227.

evaluate!(cvi)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/rCIP.jl:229.

ClusterValidityIndices.get_internal_label! — Method

get_internal_label!(
    label_map::Dict{Int64, Int64},
    label::Integer
) -> Integer

Summary

Get the internal label and update the label map if the label is new.

Arguments

label_map::LabelMap: label map to extract the internal label from.

label::Integer: the external label that corresponds to an internal label.

Method List / Definition Locations

get_internal_label!(label_map, label)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/common.jl:163.

ClusterValidityIndices.param_batch! — Method

Summary

Compute the CVI parameters in batch.

This method updates only the internal parameters of the CVI algorithm in batch. When the criterion value itself is needed, use evaluate! and extract it from cvi.criterion_value.

Arguments

cvi::CVI: the stateful information of the CVI/ICVI algorithm.

data::RealMatrix: a matrix of data where rows are features and columns are samples, used in the external clustering algorithm.

labels::IntegerVector: a vector of labels that the external clustering algorithm prescribed to each column in data.

Examples

julia> my_cvi = CH()
julia> data = load_some_data()
julia> labels = my_cluster_algorithm(data)
julia> param_batch!(my_cvi, data, labels)

Method List / Definition Locations

param_batch!(cvi, data, labels)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/CH.jl:157.

param_batch!(cvi, data, labels)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/DB.jl:179.

param_batch!(cvi, data, labels)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/GD43.jl:182.

param_batch!(cvi, data, labels)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/GD53.jl:179.

param_batch!(cvi, data, labels)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/PS.jl:138.

param_batch!(cvi, data, labels)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/WB.jl:164.

param_batch!(cvi, data, labels)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/XB.jl:175.

param_batch!(cvi, data, labels)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/cSIL.jl:200.

param_batch!(cvi, data, labels)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/rCIP.jl:183.

ClusterValidityIndices.param_inc! — Method

Summary

Compute the CVI parameters incrementally.

This method updates only internal parameters of the ICVI algorithm incrementally. When the criterion value itself is needed, use evaluate! and extract it from cvi.criterion_value.

Arguments

cvi::CVI: the stateful information of the CVI/ICVI algorithm.

sample::RealVector: a vector of features used in the external clustering algorithm.

label::Integer: the label that the external clustering algorithm prescribed to the sample.

Examples

julia> my_cvi = CH()
julia> data = load_some_data()
julia> labels = my_cluster_algorithm(data)
julia> param_inc!(my_cvi, data[:, 1], labels[1])

Method List / Definition Locations

param_inc!(cvi, sample, label)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/CH.jl:99.

param_inc!(cvi, sample, label)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/DB.jl:97.

param_inc!(cvi, sample, label)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/GD43.jl:102.

param_inc!(cvi, sample, label)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/GD53.jl:102.

param_inc!(cvi, sample, label)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/PS.jl:85.

param_inc!(cvi, sample, label)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/WB.jl:104.

param_inc!(cvi, sample, label)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/XB.jl:95.

param_inc!(cvi, sample, label)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/cSIL.jl:92.

param_inc!(cvi, sample, label)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/rCIP.jl:106.

ClusterValidityIndices.setup! — Method

Summary

Internal method, sets up the CVI based upon the type of the provided sample.

Arguments

cvi::CVI: the CVI to setup to the correct dimensions.

sample::RealVector: The sample to use as a basis for setting up the CVI.

Method List / Definition Locations

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/CH.jl:89.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/DB.jl:86.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/GD43.jl:92.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/GD53.jl:92.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/PS.jl:77.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/WB.jl:94.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/XB.jl:86.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/cSIL.jl:82.

setup!(cvi, sample)

defined at /home/runner/work/ClusterValidityIndices.jl/ClusterValidityIndices.jl/src/CVI/rCIP.jl:92.