Index

CVI.DB — Type

DB

The stateful information of the Xie-Beni CVI.

CVI.DB — Method

DB()

Default constructor for the Xie-Beni (DB) Cluster Validity Index.

CVI.PS — Type

PS

The stateful information of the Xie-Beni CVI.

CVI.PS — Method

PS()

Default constructor for the Partition Separation (PS) Cluster Validity Index.

CVI.XB — Type

XB

The stateful information of the Xie-Beni CVI.

CVI.XB — Method

XB()

Default constructor for the Xie-Beni (XB) Cluster Validity Index.

CVI.evaluate! — Method

evaluate!(cvi::DB)

Compute the criterion value of the DB CVI.

CVI.evaluate! — Method

evaluate!(cvi::PS)

Compute the criterion value of the PS CVI.

CVI.evaluate! — Method

evaluate!(cvi::XB)

Compute the criterion value of the XB CVI.

CVI.expand_array — Method

expand_array(arr::Array{Float64, 1} ; n::Int64 = 1)

Expand a 1-D array by n zeros (default is 1).

CVI.expand_array — Method

expand_array(arr::Array{Float64, 2} ; n::Int64 = 1)

Expand the dimension of a 2-D array by n in each dimension (default 1).

CVI.get_cvi! — Method

get_cvi!(cvi::T, x::Array{N, 2}, y::Array{M, 1}) where {T<:AbstractCVI, N<:Real, M<:Int}

Porcelain: update compute the criterion value in batch and return it.

CVI.get_icvi! — Method

get_icvi!(cvi::T, x::Array{N, 1}, y::M) where {T<:AbstractCVI, N<:Real, M<:Int}

Porcelain: update and compute the criterion value incrementally and return it.

CVI.param_batch! — Method

param_batch!(cvi::DB, data::Array{T, 2}, labels::Array{I, 1}) where {T<:Real, I<:Int}

Compute the DB CVI in batch.

CVI.param_batch! — Method

param_batch!(cvi::PS, data::Array{T, 2}, labels::Array{I, 1}) where {T<:Real, I<:Int}

Compute the PS CVI in batch.

CVI.param_batch! — Method

param_batch!(cvi::XB, data::Array{T, 2}, labels::Array{I, 1}) where {T<:Real, I<:Int}

Compute the XB CVI in batch.

CVI.param_inc! — Method

param_inc!(cvi::DB, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}

Compute the DB CVI incrementally.

CVI.param_inc! — Method

param_inc!(cvi::PS, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}

Compute the PS CVI incrementally.

CVI.param_inc! — Method

param_inc!(cvi::XB, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}

Compute the XB CVI incrementally.