Index

ClusterValidityIndices.CH — Type

CH

The stateful information of the Calinski-Harabasz (CH) CVI.

ClusterValidityIndices.CH — Method

CH()

Default constructor for the Calinski-Harabasz (CH) Cluster Validity Index.

ClusterValidityIndices.DB — Type

DB

The stateful information of the Davies-Bouldin (DB) CVI.

ClusterValidityIndices.DB — Method

DB()

Default constructor for the Davies-Bouldin (DB) Cluster Validity Index.

ClusterValidityIndices.GD43 — Type

GD43

The stateful information of the Generalized Dunn's Index 43 (GD43) CVI.

ClusterValidityIndices.GD43 — Method

GD43()

Default constructor for the Generalized Dunn's Index 43 (GD43) Cluster Validity Index.

ClusterValidityIndices.GD53 — Type

GD53

The stateful information of the Generalized Dunn's Index 53 (GD53) CVI.

ClusterValidityIndices.GD53 — Method

GD53()

Default constructor for the Generalized Dunn's Index 53 (GD53) Cluster Validity Index.

ClusterValidityIndices.PS — Type

PS

The stateful information of the Partition Separation (PS) CVI.

ClusterValidityIndices.PS — Method

PS()

Default constructor for the Partition Separation (PS) Cluster Validity Index.

ClusterValidityIndices.WB — Type

WB

The stateful information of the WB-Index (WB) CVI.

ClusterValidityIndices.WB — Method

WB()

Default constructor for the WB-Index (WB) Cluster Validity Index.

ClusterValidityIndices.XB — Type

XB

The stateful information of the Xie-Beni (XB) CVI.

ClusterValidityIndices.XB — Method

XB()

Default constructor for the Xie-Beni (XB) Cluster Validity Index.

ClusterValidityIndices.cSIL — Type

cSIL

The stateful information of the Centroid-based Silhouette (cSIL) CVI.

ClusterValidityIndices.cSIL — Method

cSIL()

Default constructor for the Centroid-based Silhouette (cSIL) Cluster Validity Index.

ClusterValidityIndices.rCIP — Type

rCIP

The stateful information of the (Renyi's) representative Cross Information Potential (rCIP) CVI.

ClusterValidityIndices.rCIP — Method

rCIP()

Default constructor for the (Renyi's) representative Cross Information Potential (rCIP) Cluster Validity Index.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::CH)

Compute the criterion value of the Calinski-Harabasz (CH) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::DB)

Compute the criterion value of the Davies-Bouldin (DB) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::GD43)

Compute the criterion value of the Generalized Dunn's Index 43 (GD43) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::GD53)

Compute the criterion value of the Generalized Dunn's Index 53 (GD53) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::PS)

Compute the criterion value of the Partition Separation (PS) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::WB)

Compute the criterion value of the WB-Index (WB) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::XB)

Compute the criterion value of the Xie-Beni (XB) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::cSIL)

Compute the criterion value of the Centroid-based Silhouette (cSIL) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::rCIP)

Compute the criterion value of the (Renyi's) representative Cross Information Potential (rCIP) CVI.

ClusterValidityIndices.get_cvi! — Method

get_cvi!(cvi::T, x::Array{N, 2}, y::Array{M, 1}) where {T<:AbstractCVI, N<:Real, M<:Int}

Porcelain: update compute the criterion value in batch and return it.

ClusterValidityIndices.get_icvi! — Method

get_icvi!(cvi::T, x::Array{N, 1}, y::M) where {T<:AbstractCVI, N<:Real, M<:Int}

Porcelain: update and compute the criterion value incrementally and return it.

ClusterValidityIndices.get_internal_label! — Method

get_internal_label!(label_map::LabelMap, label::Int)

Get the internal label and update the label map if the label is new.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::CH, data::RealMatrix, labels::IntegerVector)

Compute the Calinski-Harabasz (CH) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::DB, data::RealMatrix, labels::IntegerVector)

Compute the Davies-Bouldin (DB) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::GD43, data::RealMatrix, labels::IntegerVector)

Compute the Generalized Dunn's Index 43 (GD43) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::GD53, data::RealMatrix, labels::IntegerVector)

Compute the Generalized Dunn's Index 53 (GD53) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::PS, data::RealMatrix, labels::IntegerVector)

Compute the Partition Separation (PS) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::WB, data::RealMatrix, labels::IntegerVector)

Compute the WB-Index (WB) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::XB, data::RealMatrix, labels::IntegerVector)

Compute the Xie-Beni (XB) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::cSIL, data::RealMatrix, labels::IntegerVector)

Compute the Centroid-based Silhouette (cSIL) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::rCIP, data::RealMatrix, labels::IntegerVector)

Compute the (Renyi's) representative Cross Information Potential (rCIP) CVI in batch.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::CH, sample::RealVector, label::Integer)

Compute the Calinski-Harabasz (CH) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::DB, sample::RealVector, label::Integer)

Compute the Davies-Bouldin (DB) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::GD43, sample::RealVector, label::Integer)

Compute the Generalized Dunn's Index 43 (GD43) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::GD53, sample::RealVector, label::Integer)

Compute the Generalized Dunn's Index 53 (GD53) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::PS, sample::RealVector, label::Integer)

Compute the Partition Separation (PS) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::WB, sample::RealVector, label::Integer)

Compute the WB-Index (WB) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::XB, sample::RealVector, label::Integer)

Compute the Xie-Beni (XB) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::cSIL, sample::RealVector, label::Integer)

Compute the Centroid-based Silhouette (cSIL) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::rCIP, sample::RealVector, label::Integer)

Compute the (Renyi's) representative Cross Information Potential (rCIP) CVI incrementally.

ClusterValidityIndices.relabel_cvi_data — Method

relabel_cvi_data(labels::IntegerVector)

Relabels the vector to present new labels in incremental order.

ClusterValidityIndices.setup! — Method

setup!(cvi::CH, sample::Vector{T}) where {T<:Real}

ClusterValidityIndices.setup! — Method

setup!(cvi::DB, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.setup! — Method

setup!(cvi::GD43, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.setup! — Method

setup!(cvi::GD53, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.setup! — Method

setup!(cvi::PS, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.setup! — Method

setup!(cvi::WB, sample::Vector{T}) where {T<:Real}

ClusterValidityIndices.setup! — Method

setup!(cvi::XB, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.setup! — Method

setup!(cvi::cSIL, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.setup! — Method

setup!(cvi::rCIP, sample::Vector{T}) where {T<:RealFP}

ClusterValidityIndices.sort_cvi_data — Method

sort_cvi_data(data::RealMatrix, labels::IntegerVector)

Sorts the CVI data by the label index, assuring that clusters are provided incrementally.