Index

ClusterValidityIndices.CH — Type

CH

The stateful information of the Calinski-Harabasz (CH) CVI.

ClusterValidityIndices.CH — Method

CH()

Default constructor for the Calinski-Harabasz (CH) Cluster Validity Index.

ClusterValidityIndices.DB — Type

DB

The stateful information of the Davies-Bouldin (DB) CVI.

ClusterValidityIndices.DB — Method

DB()

Default constructor for the Davies-Bouldin (DB) Cluster Validity Index.

ClusterValidityIndices.GD43 — Type

GD43

The stateful information of the Generalized Dunn's Index 43 (GD43) CVI.

ClusterValidityIndices.GD43 — Method

GD43()

Default constructor for the Generalized Dunn's Index 43 (GD43) Cluster Validity Index.

ClusterValidityIndices.GD53 — Type

GD53

The stateful information of the Generalized Dunn's Index 53 (GD53) CVI.

ClusterValidityIndices.GD53 — Method

GD53()

Default constructor for the Generalized Dunn's Index 53 (GD53) Cluster Validity Index.

ClusterValidityIndices.PS — Type

PS

The stateful information of the Partition Separation (PS) CVI.

ClusterValidityIndices.PS — Method

PS()

Default constructor for the Partition Separation (PS) Cluster Validity Index.

ClusterValidityIndices.WB — Type

WB

The stateful information of the WB-Index (WB) CVI.

ClusterValidityIndices.WB — Method

WB()

Default constructor for the WB-Index (WB) Cluster Validity Index.

ClusterValidityIndices.XB — Type

XB

The stateful information of the Xie-Beni (XB) CVI.

ClusterValidityIndices.XB — Method

XB()

Default constructor for the Xie-Beni (XB) Cluster Validity Index.

ClusterValidityIndices.cSIL — Type

cSIL

The stateful information of the Centroid-based Silhouette (cSIL) CVI.

ClusterValidityIndices.cSIL — Method

cSIL()

Default constructor for the Centroid-based Silhouette (cSIL) Cluster Validity Index.

ClusterValidityIndices.rCIP — Type

rCIP

The stateful information of the (Renyi's) representative Cross Information Potential (rCIP) CVI.

ClusterValidityIndices.rCIP — Method

rCIP()

Default constructor for the (Renyi's) representative Cross Information Potential (rCIP) Cluster Validity Index.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::CH)

Compute the criterion value of the Calinski-Harabasz (CH) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::DB)

Compute the criterion value of the Davies-Bouldin (DB) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::GD43)

Compute the criterion value of the Generalized Dunn's Index 43 (GD43) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::GD53)

Compute the criterion value of the Generalized Dunn's Index 53 (GD53) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::PS)

Compute the criterion value of the Partition Separation (PS) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::WB)

Compute the criterion value of the WB-Index (WB) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::XB)

Compute the criterion value of the Xie-Beni (XB) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::cSIL)

Compute the criterion value of the Centroid-based Silhouette (cSIL) CVI.

ClusterValidityIndices.evaluate! — Method

evaluate!(cvi::rCIP)

Compute the criterion value of the (Renyi's) representative Cross Information Potential (rCIP) CVI.

ClusterValidityIndices.expand_array — Method

expand_array(arr::Array{Float64, 1} ; n::Int64 = 1)

Expand a 1-D array by n zeros (default is 1).

ClusterValidityIndices.expand_array — Method

expand_array(arr::Array{Float64, 2} ; n::Int64 = 1)

Expand the dimension of a 2-D array by n in each dimension (default 1).

ClusterValidityIndices.get_cvi! — Method

get_cvi!(cvi::T, x::Array{N, 2}, y::Array{M, 1}) where {T<:AbstractCVI, N<:Real, M<:Int}

Porcelain: update compute the criterion value in batch and return it.

ClusterValidityIndices.get_icvi! — Method

get_icvi!(cvi::T, x::Array{N, 1}, y::M) where {T<:AbstractCVI, N<:Real, M<:Int}

Porcelain: update and compute the criterion value incrementally and return it.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::CH, data::Array{T, 2}, labels::Array{I, 1}) where {T<:Real, I<:Int}

Compute the Calinski-Harabasz (CH) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::DB, data::Array{T, 2}, labels::Array{I, 1}) where {T<:Real, I<:Int}

Compute the Davies-Bouldin (DB) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::GD43, data::Array{T, 2}, labels::Array{I, 1}) where {T<:Real, I<:Int}

Compute the Generalized Dunn's Index 43 (GD43) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::GD53, data::Array{T, 2}, labels::Array{I, 1}) where {T<:Real, I<:Int}

Compute the Generalized Dunn's Index 53 (GD53) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::PS, data::Array{T, 2}, labels::Array{I, 1}) where {T<:Real, I<:Int}

Compute the Partition Separation (PS) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::WB, data::Array{T, 2}, labels::Array{I, 1}) where {T<:Real, I<:Int}

Compute the WB-Index (WB) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::XB, data::Array{T, 2}, labels::Array{I, 1}) where {T<:Real, I<:Int}

Compute the Xie-Beni (XB) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::cSIL, data::Array{T, 2}, labels::Array{I, 1}) where {T<:Real, I<:Int}

Compute the Centroid-based Silhouette (cSIL) CVI in batch.

ClusterValidityIndices.param_batch! — Method

param_batch!(cvi::rCIP, data::Array{T, 2}, labels::Array{I, 1}) where {T<:Real, I<:Int}

Compute the (Renyi's) representative Cross Information Potential (rCIP) CVI in batch.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::CH, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}

Compute the Calinski-Harabasz (CH) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::DB, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}

Compute the Davies-Bouldin (DB) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::GD43, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}

Compute the Generalized Dunn's Index 43 (GD43) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::GD53, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}

Compute the Generalized Dunn's Index 53 (GD53) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::PS, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}

Compute the Partition Separation (PS) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::WB, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}

Compute the WB-Index (WB) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::XB, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}

Compute the Xie-Beni (XB) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::cSIL, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}

Compute the Centroid-based Silhouette (cSIL) CVI incrementally.

ClusterValidityIndices.param_inc! — Method

param_inc!(cvi::rCIP, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}

Compute the (Renyi's) representative Cross Information Potential (rCIP) CVI incrementally.